+++ Python_3/DipAndAzimuth/ex_weighted_phase3_st_dip.py
@@ -5,7 +5,6 @@
 #
 import sys,os
 import numpy as np
-from scipy.signal import hilbert
 
 #
 # Import the module with the I/O scaffolding of the External Attribute
@@ -20,9 +19,11 @@ import extlib as xl
 xa.params = {
 	'Inputs': ['Input'],
 	'Output': ['Crl_dip', 'Inl_dip', 'True Dip', 'Dip Azimuth', 'Cplane'],
-	'ZSampMargin' : {'Value':[-2,2], 'Minimum':[-2,2], 'Symmetric': True},
+	'ZSampMargin' : {'Value':[-15,15], 'Minimum':[-2,2], 'Symmetric': True},
 	'StepOut' : {'Value': [2,2], 'Minimum': [2,2], 'Symmetric': True},
-	'Help': 'http://waynegm.github.io/OpendTect-Plugin-Docs/External_Attributes/DipandAzimuth/'
+	'Par_0' : {'Name': 'Tensor ZStepOut', 'Value': 1},
+	'Par_1' : {'Name': 'Band', 'Value': 0.9},
+	'Help': 'http://waynegm.github.io/OpendTect-Plugin-Docs/external_attributes/DipandAzimuth.html'
 }
 #
 # Define the compute function
@@ -30,29 +31,38 @@ xa.params = {
 def doCompute():
 	xs = xa.SI['nrinl']
 	ys = xa.SI['nrcrl']
-	zs = xa.params['ZSampMargin']['Value'][1] - xa.params['ZSampMargin']['Value'][0] + 1
+	zs = min(2*int(xa.params['Par_0']['Value'])+1,3)
 	kernel = xl.getGaussian(xs-2, ys-2, zs-2)
 	hxs = xs//2
 	hys = ys//2
 	inlFactor = xa.SI['zstep']/xa.SI['inldist'] * xa.SI['dipFactor']
 	crlFactor = xa.SI['zstep']/xa.SI['crldist'] * xa.SI['dipFactor']
+	N = xa.params['ZSampMargin']['Value'][1]
+	band = xa.params['Par_1']['Value']
+	hilbkernel = xl.hilbert_kernel(N, band)
 	while True:
 		xa.doInput()
 
-		s = xa.Input['Input']
-		sh = np.imag( hilbert(s) ) 
+		indata = xa.Input['Input']
+#
+#	Analytic Signal
+#
+		ansig = np.apply_along_axis(np.convolve,-1, indata, hilbkernel, mode="same")
+		sr = np.real(ansig)
+		si = np.imag(ansig)
 #
 #	Compute partial derivatives
-		sx = xl.kroon3( s, axis=0 )
-		sy = xl.kroon3( s, axis=1 )
-		sz = xl.kroon3( s, axis=2 )
-		shx = xl.kroon3( sh, axis=0 )
-		shy = xl.kroon3( sh, axis=1 )
-		shz = xl.kroon3( sh, axis=2 )
+#
+		sx = xl.kroon3( sr, axis=0 )
+		sy = xl.kroon3( sr, axis=1 )
+		sz = xl.kroon3( sr, axis=2 )
+		shx = xl.kroon3( si, axis=0 )
+		shy = xl.kroon3( si, axis=1 )
+		shz = xl.kroon3( si, axis=2 )
 		
-		px = s[1:xs-1,1:ys-1,:] * shx[1:xs-1,1:ys-1,:] - sh[1:xs-1,1:ys-1,:] * sx[1:xs-1,1:ys-1,:]
-		py = s[1:xs-1,1:ys-1,:] * shy[1:xs-1,1:ys-1,:] - sh[1:xs-1,1:ys-1,:] * sy[1:xs-1,1:ys-1,:]
-		pz = s[1:xs-1,1:ys-1,:] * shz[1:xs-1,1:ys-1,:] - sh[1:xs-1,1:ys-1,:] * sz[1:xs-1,1:ys-1,:]
+		px = sr[1:xs-1,1:ys-1,:] * shx[1:xs-1,1:ys-1,:] - si[1:xs-1,1:ys-1,:] * sx[1:xs-1,1:ys-1,:]
+		py = sr[1:xs-1,1:ys-1,:] * shy[1:xs-1,1:ys-1,:] - si[1:xs-1,1:ys-1,:] * sy[1:xs-1,1:ys-1,:]
+		pz = sr[1:xs-1,1:ys-1,:] * shz[1:xs-1,1:ys-1,:] - si[1:xs-1,1:ys-1,:] * sz[1:xs-1,1:ys-1,:]
 #
 #	Inner product of gradients
 		px2 = px * px