+++ Python_3/Experimental/LocalPolynomialApproximation/ex_lpa_tensor_dip.py
@@ -20,6 +20,7 @@ xa.params = {
 	'ZSampMargin' : {'Value':[-1,1], 'Symmetric': True},
 	'StepOut' : {'Value': [1,1]},
 	'Par_0': {'Name': 'Weight Factor', 'Value': 0.2},
+	'Select': {'Name': 'Normalzation', 'Values': ['Standard','First','Second','MinMax','Dynamic','Covariance'], 'Selection': 0},
 	'Help': 'http://waynegm.github.io/OpendTect-Plugin-Docs/External_Attributes/DipAndAzimuth/'
 }
 #
@@ -30,8 +31,9 @@ def doCompute():
 	ys = xa.SI['nrcrl']
 	zs = xa.params['ZSampMargin']['Value'][1] - xa.params['ZSampMargin']['Value'][0] + 1
 	wf = xa.params['Par_0']['Value']
+	norm = xa.params['Select']['Selection']
 	kernel = lpa3D_init(xs, ys, zs, wf)
-	gam = 1/(8*((min(xs,ys,zs)-1)*wf)**2)
+	gam = 1/(4*((min(xs,ys,zs)-1)*wf)**2)
 	inlFactor = xa.SI['zstep']/xa.SI['inldist'] * xa.SI['dipFactor']
 	crlFactor = xa.SI['zstep']/xa.SI['crldist'] * xa.SI['dipFactor']
 	while True:
@@ -39,11 +41,40 @@ def doCompute():
 		r = np.zeros((10,xa.TI['nrsamp']))
 		for i in range(0,10):
 			r[i,:] = xl.sconvolve(xa.Input,kernel[i])
-		A = np.rollaxis(np.array([[r[4],r[7]/2, r[8]/2], [r[7]/2, r[5], r[9]/2], [r[8]/2, r[9]/2, r[6]]]),2)
+#		A = np.rollaxis(np.array([[r[4],r[7]/2, r[8]/2], [r[7]/2, r[5], r[9]/2], [r[8]/2, r[9]/2, r[6]]]),2)
+		A = np.rollaxis(np.array([[r[4]*2,r[7], r[8]], [r[7], r[5]*2, r[9]], [r[8], r[9], r[6]*2]]),2)
+#		A = np.rollaxis(np.array([[2*r[4]],[2*r[5]],[2*r[6]]]),2)
 		AAT = np.einsum('...ij,...jk->...ik', A, np.swapaxes(A,1,2))
 		B = np.rollaxis(np.array([[r[1]],[r[2]],[r[3]]]),2)
 		BBT = np.einsum('...ij,...jk->...ik', B, np.swapaxes(B,1,2))
-		T = AAT+gam*BBT
+		T = np.zeros_like(BBT)
+		Gtmp = np.zeros(xa.TI['nrsamp'])
+		if (norm == 0):
+			T = AAT+gam*BBT
+		elif (norm == 1):
+			T = BBT
+		elif (norm == 2):
+			T = AAT
+		elif (norm == 3):
+#			G = np.sqrt(r[1]*r[1]+r[2]*r[2]+r[3]*r[3])
+			G = r[1]*r[1]+r[2]*r[2]+r[3]*r[3]
+			Gmin = np.amin(G)
+			Gmax = np.amax(G)
+			Gtmp = ((G-Gmin)/(Gmax-Gmin))
+			G = Gtmp[:,np.newaxis,np.newaxis]
+			T = (1-G)*AAT+ G*BBT
+		elif (norm == 4):
+#			G = np.sqrt(r[1]*r[1]+r[2]*r[2]+r[3]*r[3])
+			G = r[1]*r[1]+r[2]*r[2]+r[3]*r[3]
+			Groll = xl.rolling_window(G,51)
+			Gmin = np.pad(np.amin(Groll,-1),(25,25),'edge')
+			Gmax = np.pad(np.amax(Groll,-1),(25,25),'edge')
+			Gtmp = ((G-Gmin)/(Gmax-Gmin))
+			G = Gtmp[:,np.newaxis,np.newaxis]
+			T = (1-G)*AAT + G*BBT
+		else:
+			Droll = xl.rolling_window(xa.Input, zs)
+			T = np.pad(np.cov(Droll),(xa.params['ZSampMargin']['Value'][1],xa.params['ZSampMargin']['Value'][1]),'edge')
 		evals, evecs = np.linalg.eigh(T)
 		ndx = evals.argsort()
 		evecs = evecs[np.arange(0,T.shape[0],1),:,ndx[:,2]]
@@ -55,7 +86,8 @@ def doCompute():
 		xa.Output['True Dip'] = np.sqrt(xa.Output['Crl_dip']*xa.Output['Crl_dip']+xa.Output['Inl_dip']*xa.Output['Inl_dip'])
 		xa.Output['Dip Azimuth'] = np.degrees(np.arctan2(xa.Output['Inl_dip'],xa.Output['Crl_dip']))
 		coh = (eval2-eval1)/(eval2+eval1) 
-		xa.Output['Coherency'] = coh * coh
+#		xa.Output['Coherency'] = coh * coh
+		xa.Output['Coherency'] = Gtmp
 
 		xa.doOutput()
 #