+++ CHANGES.md
@@ -1,4 +1,9 @@
 # Changes
+## Jan, 2018
+ - Add ex_addnoise.py : add noise to data
+ - Add ex_dip.py :  calculate true (polar) dip and dip azimuth from inline and crossline dip
+ - Change ex_\*phase3\*_dip.py : to use convolutional hilbert transform operator
+ 
 ## Dec, 2016
  - Add Python scripts for lowpass, highpass, bandpass and band reject spatial (k-k) filter with circular (ex_spatial_filter_circular.py) and rectangular symmetry (ex_spatial_filter_rectangular.py)